+++ src/camillafir/dsp/camillafir_leveling.py
@@ -12,7 +12,6 @@ Note
 """
 
 from __future__ import annotations
-from typing import Optional
 from typing import Tuple
 import numpy as np